*Update table storage endpoint *ETag set property

+++ AzureStorageLibrary/Model/Product.cs
@@ -5,7 +5,7 @@ namespace AzureStorageLibrary.Model
     public class Product : TableEntity
     {
         public string Name { get; set; }
-        public double Price { get; set; }
+        public long Price { get; set; }
         public int Stock { get; set; }
         public string Color { get; set; }
     }
